refactor(Parser): 컨트롤러 패키지로 이동
